Refactor Identity
